refactor: #98 - Separa `SimpleCheckbox` em `Checkbox` e `LabeledCheckbox`
